rendering: fix cryptic error messages on missing components
Previously, the `render_component` would crash with an exception, stating
`UnknownDependencyError: Undefined -> args`, if a component name was not found
in the template context.
This happened because Jinja2 returns `Undefined` if you try to look up an
undefined key in a template context. Since the `Undefined` type is a class,
therefore callable, falk thinks its a component, depending on `args`.

This patch fixes this issue by updating the pyx transpiling code to not inject
the component itself, but its name. The name is then resolved in the
`_render_component`, provoking a proper key error, resulting in a user
friendly `UnknownComponentError`.

Signed-off-by: Florian Scherf <mail@florianscherf.de>