rendering: render_component: disable dependency caching
Currently, it is not possible to flag providers as uncacheable which becomes
a problem for providers that depend on dependencies like `caller` because
they need to be reevaluated on every call to `run_callback`, not only for
every request.
Since we don't have a solution for this yet, this patch disables dependency
caching entirely for now.

Signed-off-by: Florian Scherf <mail@florianscherf.de>