request handling: refactor mutation request parsing
This patch moves the information whether a request is a mutation request into
a flag on the request itself. This reduces the logic in
`falk.request_handling.handle_request` which is supposed to be customizable.

This patch also adds `requestType: "falk/mutation"` to the JSON body of
mutation request to make them easier to detect.

Signed-off-by: Florian Scherf <mail@florianscherf.de>