rendering: render_component: only generate tokens if necessary
Previously, `render_component` would generate a new token on every run, even
though not every run would yield a render able component. For example, if an
component has another component as its root like this:

    def ExampleComponent(context):
        from falk.components import HTML5Base

        context.update({
            "HTML5Base": HTML5Base,
        })

        return """
            <HTML5Base title="Hello World">
                <h1>Hello World</h1>
            </HTML5Base>
        """

This code should result in one render able component (`ExampleComponent`),
just using `HTML5Base`, with only one token.
The old code would generate a token for `ExampleComponent` and `HTML5Base`.

This patch fixes this issue by checking if there is already a generated token
for the current component and if so just pass it down the pipe.

Signed-off-by: Florian Scherf <mail@florianscherf.de>