client: runHook: add error handling
Previously, when running a hook like `render`, `runHook` would crash and stop
on the first node that raises an exception. Since it is end user code that we
are running here, we need to expect every line to crash.

This patch wraps the call to the end user code into a try catch statement and
logs errors to the console.
This is no actual handling, but at least the chain doesn't stop on the first
error anymore.

Signed-off-by: Florian Scherf <mail@florianscherf.de>