add support for file uploads
This patch adds support for multipart file uploads using starlette. The code
is contained in a new module called `asgi2` because there are still problems
with the starlette code. Starlette was originally chosen so we don't have to
deal with multipart ourself when it comes to file and memory management and
file quotas. The starlette implementation is very opinionated and the quotas
don't support all features that we would want.

Future experiments with python-multipart will tell which implementation
will stay.

Signed-off-by: Florian Scherf <mail@florianscherf.de>